# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: EdgeAI Control Plane
  version: 0.1.0
  description: Profile → Device/Node → Workflow DAG → Run/Task/Attempt와 검증된 결과 조회 API입니다.
    실제 Kubernetes 실행은 서버의 runtime 설정이 활성화된 환경에서 동작합니다. 쓰기 요청은 실제 DB에 반영되고 Swagger의 CSRF 토큰/쿠키는 자동 연결됩니다.
    복원 DB의 조회 전용 점검 모드에서는 실행 기능이 비활성화되고 CSRF가 유효한 쓰기 요청도 403으로 거절됩니다.
    일반 관리 API의 변경 요청은 먼저 감사 접수 기록을 저장하며 X-EdgeAI-Audit-Id를 반환합니다. 접수 저장 실패는503/AUDIT_STORE_UNAVAILABLE이며 변경을 실행하지 않습니다.
    감사의 HTTP 결과는 비동기 작업 성공을 뜻하지 않으며, 결과가 없는 접수 기록은 OUTCOME_UNKNOWN입니다. 복구 점검 모드에서는 감사 쓰기도 비활성화됩니다.
servers:
- url: /
security: []
paths:
...
  /api/v1/profiles/{kind}:
    parameters:
    - $ref: '#/components/parameters/Kind'
    post:
      operationId: publishProfile
      summary: 프로필 새 버전 등록
      description: 'DEVICE(장치 규격), SERVICE(서비스 규격), VD(가상 장치 규격) 중 종류를 선택해 JSON 문서를 발행합니다.


        새 key/version은 201, 같은 종류·key·version에 같은 내용을 재전송하면 기존 id와 생성 시각을 유지한 200을 반환합니다. 같은 버전에 다른 내용을 보내면 409이며 version을
        올려 등록해야 합니다. 발행 후 수정·삭제할 수 없습니다.


        세션 쿠키·CSRF 토큰이 필요합니다. 요청은 UTF-8 기준 64 KiB 이하, spec은 비어 있지 않은 JSON 객체여야 합니다. 최대 중첩 깊이는 32이며 중복 JSON 키와 key/version/spec
        이외의 최상위 필드는 거절합니다. 아래 규격은 입력 예시이며 실제 장치/런타임 호환성 검증은 후속 개발 범위입니다.'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishProfile'
            example:
              key: temperature-sensor
              version: 1.0.0
              spec:
                protocol: mqtt
                unit: celsius
      responses:
        '201':
          description: 새 버전을 발행했습니다. Location 헤더로 상세 조회할 수 있습니다.
          headers:
            Location:
              schema:
                type: string
              description: Version lookup path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileVersion'
        '200':
          description: 동일 내용의 기존 버전입니다. 중복 생성하지 않고 기존 결과를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileVersion'
        '400':
          $ref: '#/components/responses/Invalid'
        '403':
          description: Missing or invalid CSRF token
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: 요청 또는 정규화된 규격이 UTF-8 64 KiB를 초과했습니다.
        '503':
          $ref: '#/components/responses/Unavailable'
      tags:
      - Profile
    get:
      operationId: listProfiles
      summary: 종류별 프로필 버전 목록 조회
      parameters:
      - in: query
        name: key
        description: 프로필 키 정확 일치 필터. 생략하면 선택한 종류의 전체 키를 조회합니다.
        schema:
          $ref: '#/components/schemas/ProfileKey'
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: 한 페이지의 최대 항목 수. 1~100, 기본 20입니다.
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          maximum: 1000000
          default: 0
        description: 건너뛸 항목 수. 이전 응답의 nextOffset을 넣으면 다음 페이지입니다.
      responses:
        '200':
          description: 발행 버전 목록. items가 빈 배열이면 결과가 없으며 nextOffset이 null이면 마지막 페이지입니다.
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - nextOffset
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileVersion'
                  nextOffset:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/Invalid'
        '503':
          $ref: '#/components/responses/Unavailable'
      tags:
      - Profile
      description: '선택한 종류의 발행 버전을 페이지로 조회합니다. key를 지정하면 해당 키와 정확히 일치하는 버전만 반환합니다.


        key와 version의 문자열 오름차순입니다. 따라서 1.0.0 → 10.0.0 → 2.0.0 순서일 수 있으며 최신 버전을 계산하는 API가 아닙니다. 다음 페이지는 nextOffset을 offset으로
        전달하며, null이면 마지막입니다.'
  /api/v1/profiles/{kind}/{key}/versions/{version}:
    parameters:
    - $ref: '#/components/parameters/Kind'
    - in: path
      name: key
      required: true
      schema:
        $ref: '#/components/schemas/ProfileKey'
      description: 조회할 프로필의 고유 키입니다.
    - in: path
      name: version
      required: true
      schema:
        $ref: '#/components/schemas/Version'
      description: '조회할 정확한 발행 버전입니다. 예: 1.0.0'
    get:
      operationId: getProfileVersion
      summary: 프로필의 특정 버전 상세 조회
      responses:
        '200':
          description: Published version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileVersion'
        '400':
          $ref: '#/components/responses/Invalid'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
      tags:
      - Profile
      description: kind·key·version으로 발행 버전 하나를 정확히 조회합니다. 저장된 spec, 내용 식별용 SHA-256 digest, 최초 생성 시각을 반환합니다. 이전 버전 내용을 확인하거나
        후속 장치/실행 기능에서 참조할 버전을 선택할 때 사용합니다. 해당 버전이 없으면 404를 반환합니다.
    delete:
      operationId: deleteProfileVersion
      tags: [Profile]
      summary: 사용하지 않는 프로필 버전 삭제
      description: 장치·가상 장치·워크플로·실행 이력 또는 VD 프로필의 참조가 있으면 PROFILE_IN_USE(409)로 거절합니다. 다른 버전은 유지합니다.
      security:
      - csrfToken: []
      responses:
        '204':
          description: 버전 삭제 완료
        '400':
          $ref: '#/components/responses/Invalid'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: PROFILE_IN_USE — 다른 자원에서 참조하는 버전
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/v1/sensors/registration-options:
    get:
      operationId: getSensorRegistrationOptions
      summary: 배포된 수집기의 지원 센서 등록 항목 조회
      tags: [센서]
      responses:
        '200':
          description: 현재 EdgeX 서비스와 프로필이 확인된 지원 항목. 연결 가능 여부나 실측 수신을 보장하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationCatalog'}
        '502': {description: EdgeX 목록 조회 실패 또는 조회 한도 초과}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/registrations:
    post:
      operationId: registerSensor
      summary: 지원하는 센서를 EdgeX에 수동 등록
      tags: [센서]
      security: [{csrfToken: []}]
      description: >-
        배포된 Arduino Serial 또는 Sense HAT I2C 프로필을 사용한다. 새 드라이버나 프로필을 설치하지 않는다.
        동일 이름·연결 정보는 기존 등록을 반환하고 이름 또는 동일 포트의 측정 항목 충돌은409다.
        등록은 EdgeX DB에 저장되며 레포 YAML에는 자동 반영되지 않는다. 결과 미확인 시 자동 재시도하지 않는다.
        같은 이름·입력값을 유지한 재요청으로 확인할 수 있다. 등록 완료는 실측 수신 확인과 다르다.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SensorRegistrationRequest'}
      responses:
        '201':
          description: 새 등록 저장과 읽기 확인 완료
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationResult'}
        '200':
          description: 같은 이름·설정으로 이미 등록됨
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationResult'}
        '400': {description: 지원하지 않는 센서 또는 연결 설정}
        '403': {description: CSRF 검증 실패}
        '409': {description: 등록 충돌 또는 수집기·프로필 없음}
        '502': {description: EdgeX 조회 실패 또는 등록 결과 미확인}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/readings:
    get:
      operationId: getSensorReadings
      summary: EdgeX 센서의 최근 측정값 조회
      tags: [센서]
      parameters:
      - {name: device, in: query, required: true, schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'}}
      - {name: resource, in: query, schema: {type: string}, description: 특정 측정 항목으로 제한}
      - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
      responses:
        '200':
          description: 최근 N개 측정값을 측정 시각 오름차순으로 반환. 미수집은 빈 배열이며 0과 구분한다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorReadings'}
        '400': {description: 잘못된 조회 조건}
        '404': {description: 센서 없음}
        '502': {description: EdgeX 조회 실패}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/commands:
    get:
      operationId: getSensorCommands
      summary: 센서가 실제 지원하는 읽기·쓰기 명령 조회
      tags: [센서]
      parameters:
      - {name: device, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: Core Command에서 조회한 명령. 외부 URL이나 프로토콜 인증 정보는 반환하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorCommands'}
        '404': {description: 센서 없음}
        '502': {description: EdgeX 조회 실패}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/command:
    post:
      operationId: executeSensorCommand
      summary: EdgeX에 등록된 센서 명령 실행
      tags: [센서]
      description: CSRF가 필요하다. 명령 지원 여부와 adminState를 실행 전에 재조회한다. GET은 ds-pushevent=false로 직접 읽으며 PUT은 지정된 항목의 문자열 값을 전달한다. 자동 재시도하지 않으며 시간 초과 시 실제 장치 결과를 확인해야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SensorCommandRequest'}
      responses:
        '200':
          description: EdgeX 명령 응답. PUT 응답은 물리 장치 변경 검증을 뜻하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorCommandResult'}
        '400': {description: 지원하지 않는 명령 또는 잘못된 값}
        '403': {description: CSRF 검증 실패}
        '404': {description: 센서 또는 명령 없음}
        '409': {description: 장치 잠김}
        '502': {description: EdgeX 명령 응답 실패 또는 결과 미확인}
        '503': {description: EdgeX 비활성화}
...
components:
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 쓰기 요청은 세션 쿠키와 함께 이 헤더가 필요합니다. Swagger는 자동 발급·전달하므로 이 입력칸을 비워 두세요. 직접 호출할 때는 getCsrfToken을 먼저 사용합니다.
  parameters:
    Kind:
      in: path
      name: kind
      required: true
      schema:
        type: string
        enum:
        - DEVICE
        - SERVICE
        - VD
      description: 'DEVICE: 장치 규격 / SERVICE: 서비스 규격 / VD: 가상 장치 규격. 동일한 key/version이라도 종류가 다르면 별개입니다.'
  responses:
...
    Invalid:
      description: 종류·키·버전·페이지·JSON 형식이 잘못되었습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_PROFILE
            message: 종류, 키, 버전과 JSON 규격을 확인하세요. 규격은 비어 있지 않은 객체여야 합니다.
    Conflict:
      description: 같은 종류·키·버전에 다른 내용이 이미 발행되었습니다. 새 버전을 사용하세요.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_CONFLICT
            message: 이미 발행한 버전입니다. 내용을 바꾸려면 새 버전을 등록하세요.
    NotFound:
      description: 해당 발행 버전이 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_NOT_FOUND
            message: 해당 Profile 버전을 찾을 수 없습니다.
    Unavailable:
      description: DB 연결 등 저장소 문제로 요청을 처리할 수 없습니다. 복구 후 재시도하세요.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_STORE_UNAVAILABLE
            message: Profile 저장소에 연결할 수 없습니다. 잠시 후 다시 시도하세요.
  schemas:
    SensorRegistrationTemplate:
      type: object
      required: [id, label, serviceName, nodeName, profileName, protocol, resources, endpoint, baudRates]
      properties:
        id: {type: string}
        label: {type: string}
        serviceName: {type: string}
        nodeName: {type: string}
        profileName: {type: string}
        protocol: {type: string, enum: [SERIAL, I2C]}
        resources: {type: array, items: {type: string}}
        endpoint: {type: string}
        baudRates: {type: array, items: {type: integer, enum: [1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400, 460800, 921600]}}
    SensorRegistrationCatalog:
      type: object
      required: [templates]
      properties:
        templates: {type: array, items: {$ref: '#/components/schemas/SensorRegistrationTemplate'}}
    SensorRegistrationRequest:
      type: object
      additionalProperties: false
      required: [name, templateId, endpoint, physicalDeviceId]
      properties:
        name: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        templateId: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        endpoint: {type: string, maxLength: 80, description: 'Serial은 /dev/edgeai/ 아래의 단일 장치명, Sense HAT는 /dev/i2c-1'}
        physicalDeviceId: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        baudRate: {type: [integer, 'null'], enum: [null, 1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400, 460800, 921600]}
    SensorRegistrationResult:
      type: object
      required: [name, serviceName, profileName, created, adminState, operatingState]
      properties:
        name: {type: string}
        serviceName: {type: string}
        profileName: {type: string}
        created: {type: boolean}
        adminState: {type: string}
        operatingState: {type: string}
    SensorReading:
      type: object
      required: [resource, valueType, value, units, observedAt]
      properties:
        resource: {type: string}
        valueType: {type: string}
        value: {type: string, description: "원본 문자열 값. Binary는 [binary]로 표시"}
        units: {type: string}
        observedAt: {type: string, format: date-time}
    SensorReadings:
      type: object
      required: [device, fetchedAt, readings]
      properties:
        device: {type: string}
        fetchedAt: {type: string, format: date-time}
        readings: {type: array, items: {$ref: '#/components/schemas/SensorReading'}}
    SensorCommand:
      type: object
      required: [name, readable, writable, parameters]
      properties:
        name: {type: string}
        readable: {type: boolean}
        writable: {type: boolean}
        parameters:
          type: array
          items:
            type: object
            required: [resource, valueType]
            properties:
              resource: {type: string}
              valueType: {type: string}
    SensorCommands:
      type: object
      required: [device, adminState, operatingState, commands]
      properties:
        device: {type: string}
        adminState: {type: string}
        operatingState: {type: string}
        commands: {type: array, items: {$ref: '#/components/schemas/SensorCommand'}}
    SensorCommandRequest:
      type: object
      required: [device, command, method, values]
      properties:
        device: {type: string}
        command: {type: string}
        method: {type: string, enum: [GET, PUT]}
        values: {type: object, additionalProperties: {type: string, maxLength: 4096}}
    SensorCommandResult:
      type: object
      required: [device, command, method, completedAt, readings]
      properties:
        device: {type: string}
        command: {type: string}
        method: {type: string, enum: [GET, PUT]}
        completedAt: {type: string, format: date-time}
        readings: {type: array, items: {$ref: '#/components/schemas/SensorReading'}}
...
    ProfileKey:
      type: string
      maxLength: 100
      pattern: ^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$
      description: 소문자 영문으로 시작하는 프로필 키. 영문 소문자·숫자와 구분자 점/밑줄/하이픈을 사용하며 최대 100자입니다.
      example: temperature-sensor
    Version:
      type: string
      maxLength: 32
      pattern: ^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$
      description: MAJOR.MINOR.PATCH 형태의 숫자 버전. 01처럼 앞에 0을 붙이거나 prerelease 문자를 붙일 수 없습니다.
      example: 1.0.0
    PublishProfile:
      type: object
      additionalProperties: false
      required:
      - key
      - version
      - spec
      properties:
        key:
          $ref: '#/components/schemas/ProfileKey'
        version:
          $ref: '#/components/schemas/Version'
        spec:
          type: object
          minProperties: 1
          additionalProperties: true
          description: 비어 있지 않은 JSON 규격 문서입니다. 현재 종류별 실행 스키마·외부 참조의 호환성은 검증하지 않습니다.
    ProfileVersion:
      type: object
      required:
      - id
      - kind
      - key
      - version
      - spec
      - digest
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: 이 발행 버전의 서버 생성 UUID입니다.
        kind:
          type: string
          enum:
          - DEVICE
          - SERVICE
          - VD
          description: 프로필 종류입니다. 종류마다 키·버전 공간이 구분됩니다.
        key:
          $ref: '#/components/schemas/ProfileKey'
        version:
          $ref: '#/components/schemas/Version'
        spec:
          type: object
          additionalProperties: true
          description: 발행 시 저장된 JSON 규격입니다.
        digest:
          type: string
          pattern: ^sha256:[a-f0-9]{64}$
          description: 종류와 정규화된 spec의 SHA-256 식별값. JSON 키 순서나 동등한 숫자 표기는 같은 값이며, spec 또는 종류가 바뀌면 달라집니다. 상세 규칙은 ADR 0002를 따릅니다.
        createdAt:
          type: string
          format: date-time
          description: 최초 발행 시각입니다. 동일 내용 재등록 시 유지됩니다.
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_PROFILE
          - PROFILE_CONFLICT
          - PROFILE_NOT_FOUND
          - PROFILE_STORE_UNAVAILABLE
          - PAYLOAD_TOO_LARGE
          - INVALID_DEVICE
          - DEVICE_CONFLICT
          - DEVICE_NOT_FOUND
          - NODE_NOT_FOUND
          - NODE_NOT_READY
          - DEVICE_RELEASED
          - STALE_SESSION
          - OBSERVATION_CONFLICT
          - DEVICE_STORE_UNAVAILABLE
          - INVALID_WORKFLOW
          - WORKFLOW_NOT_FOUND
          - WORKFLOW_CONFLICT
          - RUN_NOT_FOUND
          - TASK_NOT_FOUND
          - IDEMPOTENCY_CONFLICT
          - CANNOT_CANCEL
          - STREAM_NOT_IMPLEMENTED
          - STREAM_EXECUTION_POLICY_UNSUPPORTED
          - STREAM_RECOVERY_UNSUPPORTED
          - STREAM_SOURCE_INACTIVE
          - WORKFLOW_STORE_UNAVAILABLE
          - INVALID_TASK_ID
          - RESULT_STORE_UNAVAILABLE
          - RUNTIME_DISABLED
          - REMOTE_DISABLED
          - REMOTE_PROVIDER_NOT_FOUND
          - REMOTE_CONFIGURATION_CHANGED
          - REMOTE_TELEMETRY_UNSUPPORTED
          - VD_EXECUTION_DISABLED
          - VD_NOT_READY
          - VD_SERVICE_MISMATCH
          - VD_AUTOMATIC_OFFLOAD_UNSUPPORTED
          - OFFLOAD_SOURCE_CHANGED
          - OFFLOAD_LIMIT
          - OFFLOAD_RECOVERY_UNSUPPORTED
          - OFFLOAD_TARGET_INVALID
          - OPERATION_NOT_FOUND
          - INVALID_VIRTUAL_DEVICE
          - VD_NOT_FOUND
          - VD_CONFLICT
          - VD_RELEASED
          - VD_SOURCE_INCOMPATIBLE
          - VD_STORE_UNAVAILABLE
          - DEVICE_IN_USE
        message:
          type: string
...